refactor(StationRepository): station fake 객체 리팩토링
