+++ src/main/java/nextstep/subway/station/domain/JpaStationRepository.java
@@ -1,0 +1,6 @@
+package nextstep.subway.station.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface JpaStationRepository extends StationRepository, JpaRepository<Station, Long> {
+}
