+++ src/main/java/nextstep/subway/station/domain/StationRepository.java
@@ -1,10 +1,15 @@
 package nextstep.subway.station.domain;
 
-import org.springframework.data.jpa.repository.JpaRepository;
-
 import java.util.List;
+import java.util.Optional;
 
-public interface StationRepository extends JpaRepository<Station, Long> {
-    @Override
+public interface StationRepository {
+    Optional<Station> findById(Long id);
+
     List<Station> findAll();
+
+    Station save(Station toStation);
+
+    void deleteById(Long id);
+
 }
